Add RSpec support
